ServerAccessor class is working for upload and download operations

+++ UploadUnit/src/main/java/ServerAccessor.java
@@ -8,26 +8,36 @@ import com.amazonaws.auth.EnvironmentVariableCredentialsProvider;
 import com.amazonaws.regions.Regions;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
+import com.amazonaws.services.s3.model.DeleteObjectRequest;
+import com.amazonaws.services.s3.model.GetObjectRequest;
 import com.amazonaws.services.s3.model.PutObjectRequest;
+import com.amazonaws.services.s3.model.S3Object;
+import org.apache.commons.io.FileUtils;
 
-import java.io.File;
-import java.io.IOException;
+import java.io.*;
 
 
-public class UploadObjectSingleOperation {
-    private static String bucketName     = System.getenv("BUCKET_NAME");
-    private static String keyName        = "BomBom2";
-    private static String uploadFileName = "/Users/furkansahin/CrypDist/UploadUnit/src/main/java/cs442_clock.pdf";
+public class ServerAccessor {
+    private String bucketName;
+    private AmazonS3 s3client;
 
-    public static void main(String[] args) throws IOException {
-        AmazonS3 s3client = AmazonS3ClientBuilder.standard()
+    public ServerAccessor(){
+        bucketName = System.getenv("BUCKET_NAME");
+
+        s3client = AmazonS3ClientBuilder.standard()
                 .withCredentials(new EnvironmentVariableCredentialsProvider())
                 .withRegion(Regions.EU_CENTRAL_1).build();
+    }
+
+    public void upload(String fileName, String filePath) throws Exception {
+        if (doesObjectExist(fileName))
+            throw new Exception("file already exists!");
+
         try {
             System.out.println("Uploading a new object to S3 from a file\n");
-            File file = new File(uploadFileName);
+            File file = new File(filePath);
             s3client.putObject(new PutObjectRequest(
-                    bucketName, keyName, file));
+                    bucketName, fileName, file));
 
         } catch (AmazonServiceException ase) {
             System.out.println("Caught an AmazonServiceException, which " +
@@ -47,5 +57,31 @@ public class UploadObjectSingleOperation {
                     "such as not being able to access the network.");
             System.out.println("Error Message: " + ace.getMessage());
         }
+    }
+
+    public void download(String fileName, String toDirectory) throws IOException {
+        S3Object object = s3client.getObject(new GetObjectRequest(bucketName, fileName));
+        InputStream objectData = object.getObjectContent();
+
+        byte[] buffer = new byte[(int) object.getObjectMetadata().getContentLength()];
+        objectData.read(buffer);
+
+        FileUtils.writeByteArrayToFile(new File(toDirectory), buffer);
+
+        try {
+            objectData.close();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    public boolean doesObjectExist(String fileName)
+    {
+        return s3client.doesObjectExist(bucketName, fileName);
+    }
+
+    void delete(String fileName)
+    {
+        s3client.deleteObject(new DeleteObjectRequest(bucketName, fileName));
     }
 }
